add support to install profvis when opening profiles
